+++ website/src/ts/scripts/slds-components.ts
@@ -1,5 +1,29 @@
-import '@oddbird/popover-polyfill';
+import { faGithub, faSlack } from '@fortawesome/free-brands-svg-icons';
+// eslint-disable-next-line import/order
 import '@webcomponents/scoped-custom-element-registry/scoped-custom-element-registry.min.js';
+import {
+  faArrowUpRightFromSquare,
+  faBug,
+  faCode,
+  faCopy,
+  faFileEdit,
+  faFloppyDisk,
+  faMessagesQuestion,
+  faPenToSquare,
+  faPencilRuler,
+  faTrash,
+  faTruckFast,
+  faXmark,
+  faGear as farGear
+} from '@fortawesome/pro-regular-svg-icons';
+import {
+  faEnvelope,
+  faHandBackPointUp,
+  faSchool,
+  faScreenUsers,
+  faGear as fasGear
+} from '@fortawesome/pro-solid-svg-icons';
+import '@oddbird/popover-polyfill';
 import '@sl-design-system/accordion/register.js';
 import '@sl-design-system/avatar/register.js';
 import '@sl-design-system/badge/register.js';
@@ -13,12 +37,15 @@ import '@sl-design-system/drawer/register.js';
 import '@sl-design-system/editor/register.js';
 import '@sl-design-system/form/register.js';
 import '@sl-design-system/grid/register.js';
+import { Icon } from '@sl-design-system/icon';
 import '@sl-design-system/icon/register.js';
 import '@sl-design-system/inline-message/register.js';
+import { MessageDialog } from '@sl-design-system/message-dialog';
 import '@sl-design-system/message-dialog/register.js';
 import '@sl-design-system/popover/register.js';
 import '@sl-design-system/progress-bar/register.js';
 import '@sl-design-system/radio-group/register.js';
+import { setup } from '@sl-design-system/sanoma-learning';
 import '@sl-design-system/search-field/register.js';
 import '@sl-design-system/select/register.js';
 import '@sl-design-system/skeleton/register.js';
@@ -27,26 +54,8 @@ import '@sl-design-system/switch/register.js';
 import '@sl-design-system/tabs/register.js';
 import '@sl-design-system/text-area/register.js';
 import '@sl-design-system/text-field/register.js';
+import '@sl-design-system/toggle-button/register.js';
 import '@sl-design-system/tooltip/register.js';
-import { MessageDialog } from '@sl-design-system/message-dialog';
-import { setup } from '@sl-design-system/sanoma-learning';
-import { Icon } from '@sl-design-system/icon';
-import {
-  faArrowUpRightFromSquare,
-  faBug,
-  faCode,
-  faCopy,
-  faFileEdit,
-  faFloppyDisk,
-  faMessagesQuestion,
-  faPenToSquare,
-  faPencilRuler,
-  faTrash,
-  faTruckFast,
-  faXmark
-} from '@fortawesome/pro-regular-svg-icons';
-import { faEnvelope, faHandBackPointUp, faSchool, faScreenUsers } from '@fortawesome/pro-solid-svg-icons';
-import { faGithub, faSlack } from '@fortawesome/free-brands-svg-icons';
 
 setup();
 
@@ -68,7 +77,9 @@ Icon.register(
   faSlack,
   faTrash,
   faTruckFast,
-  faXmark
+  faXmark,
+  farGear,
+  fasGear
 );
 
 declare global {